Move source connector logic from shared.go
