+++ source.go
@@ -15,7 +15,9 @@
 package activemq
 
 import (
+	"bytes"
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"sync"
@@ -162,6 +164,19 @@ func (s *Source) Read(ctx context.Context) (sdk.Record, error) {
 	}
 }
 
+// metadataFromMsg extracts all the present headers from a stomp.Message into
+// sdk.Metadata.
+func metadataFromMsg(msg *stomp.Message) sdk.Metadata {
+	metadata := make(sdk.Metadata)
+
+	for i := range msg.Header.Len() {
+		k, v := msg.Header.GetAt(i)
+		metadata[k] = v
+	}
+
+	return metadata
+}
+
 func (s *Source) Ack(ctx context.Context, position sdk.Position) error {
 	pos, err := parseSDKPosition(position)
 	if err != nil {
@@ -186,4 +201,31 @@ func (s *Source) Ack(ctx context.Context, position sdk.Position) error {
 
 func (s *Source) Teardown(ctx context.Context) error {
 	return teardown(ctx, s.subscription, s.conn)
+}
+
+type Position struct {
+	MessageID string `json:"message_id"`
+	Queue     string `json:"queue"`
+}
+
+func parseSDKPosition(sdkPos sdk.Position) (Position, error) {
+	decoder := json.NewDecoder(bytes.NewBuffer(sdkPos))
+	decoder.DisallowUnknownFields()
+
+	var p Position
+	if err := decoder.Decode(&p); err != nil {
+		return p, fmt.Errorf("failed to parse position: %w", err)
+	}
+
+	return p, nil
+}
+
+func (p Position) ToSdkPosition() sdk.Position {
+	bs, err := json.Marshal(p)
+	if err != nil {
+		// this should never happen
+		panic(err)
+	}
+
+	return sdk.Position(bs)
 }
